Made lines less broken-looking, refactored a bunch of things and moved everything to different files.
